requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
packaging==24.2
pandas==2.2.3
pillow==10.2.0
plotly==5.24.1
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
# -e git+https://github.com/facebookresearch/xformers.git@6e10bd21ac6fc878657b24684723ccd05e41d385#egg=xformers
xformers